pyproject.toml
[tool.poetry]
name = "pydantic-modelgen"
version = "0.1.0"
description = "Create Pydantic `BaseModel`s from JSON Schema at runtime."
authors = ["Niklas von Moers <niklasvmoers@protonmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "pydanticmodelgen", from = "pydantic-modelgen/src" }]

[tool.poetry.dependencies]
python = ">= 3.8"
pydantic = ">= 2.7.0"
jsonschema = ">= 4.22.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
sphinx = { version = "^7.3.7", python = ">=3.9" }
ruff = "^0.4.8"
mypy = "^1.10.0"
pytest = "^8.2.2"
types-jsonschema = "^4.22.0.20240501"
pytest-cov = "^5.0.0"
lxml = "^5.2.2"
bandit = "^1.7.8"
tox = "^4.15.1"
types-python-dateutil = "^2.9.0.20240316"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude_dirs = ["venv", ".venv"]

[tool.mypy]
disallow_any_explicit = false

[tool.pytest.ini_options]
testpaths = [
    "pydantic-modelgen/tests"
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 100
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["UP006", "UP035"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = ["UP006", "UP035"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
